tests: port test_data framework from loxigen
The XDR and binary portions of a testcase are the same for all targets. Rather
than duplicating them for each target, we use a simple file format that
contains the XDR, binary, and target-specific representation. This makes it
obvious when some targets are missing testcases that are implemented for
others.